before oapi refactor
