+++ internal/rest/router/router.go
@@ -7,6 +7,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/shulganew/GophKeeper/internal/app"
 	"github.com/shulganew/GophKeeper/internal/app/config"
+	"github.com/shulganew/GophKeeper/internal/entities"
 	"github.com/shulganew/GophKeeper/internal/rest/handler"
 	"github.com/shulganew/GophKeeper/internal/rest/middlewares"
 )
@@ -17,7 +18,7 @@ func RouteShear(conf config.Config, application *app.UseCases) (r *chi.Mux) {
 	// Send password for enctription to middlewares.
 	r.Use(func(h http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			ctx := context.WithValue(r.Context(), middlewares.CtxPassKey{}, conf.PassJWT)
+			ctx := context.WithValue(r.Context(), entities.CtxPassKey{}, conf.PassJWT)
 			h.ServeHTTP(w, r.WithContext(ctx))
 		})
 	})
